feat: use app directory
