--- src/components/layouts.tsx
@@ -1,9 +1,0 @@
-export type LayoutProps = {
-  children: React.ReactNode;
-};
-
-export const Layout: React.FC<LayoutProps> = ({ children }: LayoutProps) => {
-  return <>{children}</>;
-};
-
-export default Layout;
