--- src/pages/_app.tsx
@@ -1,8 +1,0 @@
-import '../styles/globals.css';
-import type { AppType } from 'next/dist/shared/lib/utils';
-
-const MyApp: AppType = ({ Component, pageProps }) => {
-  return <Component {...pageProps} />;
-};
-
-export default MyApp;
